backend/app/services/rain_prediction_service.py
"""
AI-powered rain prediction service using LSTM neural network
"""

import numpy as np
import pandas as pd
from datetime import datetime, timedelta
from typing import List, Dict, Any, Tuple
import logging
import joblib
import os
from pathlib import Path

from app.schemas import WeatherData, RainPrediction
from app.services.weather_service import WeatherService

logger = logging.getLogger(__name__)


class RainPredictionService:
    def __init__(self):
        self.weather_service = WeatherService()
        self.model_path = Path("models/rain_prediction_model.pkl")
        self.scaler_path = Path("models/rain_scaler.pkl")
        self.model = None
        self.scaler = None
        self.sequence_length = 24  # Use 24 hours of data for prediction
        self.feature_columns = [
            'temperature', 'humidity', 'pressure', 'wind_speed', 
            'wind_direction', 'cloud_cover', 'precipitation'
        ]
        
        # Initialize model
        self._load_or_create_model()

    def _load_or_create_model(self):
        """Load existing model or create a new one"""
        try:
            if self.model_path.exists() and self.scaler_path.exists():
                self.model = joblib.load(self.model_path)
                self.scaler = joblib.load(self.scaler_path)
                logger.info("Loaded existing rain prediction model")
            else:
                logger.info("No existing model found, will create new one")
                self.model = None
                self.scaler = None
        except Exception as e:
            logger.error(f"Error loading model: {e}")
            self.model = None
            self.scaler = None

    async def predict_rainfall(
        self, 
        latitude: float, 
        longitude: float, 
        forecast_days: int = 7
    ) -> List[RainPrediction]:
        """Predict rainfall for the next forecast_days"""
        
        # Get historical and current weather data
        historical_data = await self.weather_service.get_historical_weather(
            latitude, longitude, days_back=30
        )
        current_data = await self.weather_service.get_current_weather(
            latitude, longitude
        )
        forecast_data = await self.weather_service.get_forecast(
            latitude, longitude, days=forecast_days
        )

        # Combine all data
        all_data = historical_data + [current_data] + forecast_data
        
        # Convert to DataFrame
        df = self._weather_data_to_dataframe(all_data)
        
        # Prepare features
        features = self._prepare_features(df)
        
        # Generate predictions
        predictions = []
        
        # 24-hour prediction
        pred_24h = await self._predict_single_period(features, 24)
        predictions.append(pred_24h)
        
        # 3-day prediction
        pred_3d = await self._predict_single_period(features, 72)
        predictions.append(pred_3d)
        
        # 7-day prediction
        pred_7d = await self._predict_single_period(features, 168)
        predictions.append(pred_7d)

        return predictions

    async def _predict_single_period(
        self, 
        features: np.ndarray, 
        hours: int
    ) -> RainPrediction:
        """Predict rainfall for a specific time period"""
        
        if self.model is None:
            # Use rule-based prediction if no model available
            return self._rule_based_prediction(features, hours)
        
        try:
            # Prepare input sequence
            if len(features) < self.sequence_length:
                # Pad with zeros if not enough data
                padded_features = np.zeros((self.sequence_length, len(self.feature_columns)))
                padded_features[-len(features):] = features[-len(features):]
                features = padded_features
            else:
                features = features[-self.sequence_length:]
            
            # Reshape for model input
            X = features.reshape(1, self.sequence_length, len(self.feature_columns))
            
            # Make prediction
            prediction = self.model.predict(X)[0]
            
            # Generate time series
            time_series = self._generate_time_series(prediction, hours)
            
            return RainPrediction(
                forecast_hours=hours,
                predicted_rainfall=float(np.sum(prediction)),
                confidence=min(85.0, 60.0 + np.std(prediction) * 10),
                time_series=time_series
            )
            
        except Exception as e:
            logger.error(f"Error in model prediction: {e}")
            return self._rule_based_prediction(features, hours)

    def _rule_based_prediction(self, features: np.ndarray, hours: int) -> RainPrediction:
        """Fallback rule-based prediction when ML model is not available"""
        
        # Simple rule-based prediction based on weather patterns
        recent_data = features[-24:] if len(features) >= 24 else features
        
        # Calculate probability based on recent weather
        humidity_avg = np.mean(recent_data[:, 1])  # humidity
        pressure_avg = np.mean(recent_data[:, 2])  # pressure
        cloud_avg = np.mean(recent_data[:, 5])     # cloud cover
        recent_rain = np.mean(recent_data[:, 6])   # recent precipitation
        
        # Simple rules for rain prediction
        rain_probability = 0.0
        
        if humidity_avg > 80:
            rain_probability += 0.3
        if pressure_avg < 1010:
            rain_probability += 0.2
        if cloud_avg > 70:
            rain_probability += 0.2
        if recent_rain > 0:
            rain_probability += 0.1
        
        # Predict rainfall amount
        predicted_rainfall = rain_probability * (hours / 24) * 5.0  # Max 5mm per day
        
        # Add some randomness for realistic variation
        noise = np.random.normal(0, 0.5)
        predicted_rainfall = max(0, predicted_rainfall + noise)
        
        # Generate time series
        time_series = self._generate_time_series(
            np.full(hours, predicted_rainfall / hours), 
            hours
        )
        
        return RainPrediction(
            forecast_hours=hours,
            predicted_rainfall=float(predicted_rainfall),
            confidence=float(rain_probability * 100),
            time_series=time_series
        )

    def _weather_data_to_dataframe(self, weather_data: List[WeatherData]) -> pd.DataFrame:
        """Convert weather data to pandas DataFrame"""
        data = []
        for wd in weather_data:
            data.append({
                'timestamp': wd.timestamp,
                'temperature': wd.temperature,
                'humidity': wd.humidity,
                'pressure': wd.pressure,
                'wind_speed': wd.wind_speed,
                'wind_direction': wd.wind_direction,
                'cloud_cover': wd.cloud_cover,
                'precipitation': wd.precipitation
            })
        
        df = pd.DataFrame(data)
        df.set_index('timestamp', inplace=True)
        df.sort_index(inplace=True)
        
        return df

    def _prepare_features(self, df: pd.DataFrame) -> np.ndarray:
        """Prepare features for model input"""
        # Select and normalize features
        features = df[self.feature_columns].values
        
        if self.scaler is not None:
            features = self.scaler.transform(features)
        else:
            # Simple normalization
            features = (features - features.mean(axis=0)) / (features.std(axis=0) + 1e-8)
        
        return features

    def _generate_time_series(self, predictions: np.ndarray, hours: int) -> List[Dict]:
        """Generate hourly time series data"""
        time_series = []
        base_time = datetime.now()
        
        for i, pred in enumerate(predictions[:hours]):
            timestamp = base_time + timedelta(hours=i)
            time_series.append({
                "timestamp": timestamp.isoformat(),
                "predicted_rainfall_mm": float(pred),
                "hour": i + 1
            })
        
        return time_series

    async def train_model(self, latitude: float, longitude: float):
        """Train the LSTM model with historical data"""
        try:
            # Get historical data
            historical_data = await self.weather_service.get_historical_weather(
                latitude, longitude, days_back=90
            )
            
            # Convert to DataFrame
            df = self._weather_data_to_dataframe(historical_data)
            
            # Prepare training data
            features = self._prepare_features(df)
            targets = df['precipitation'].values
            
            # Create sequences
            X, y = self._create_sequences(features, targets)
            
            # Train a simple model (in production, use proper LSTM)
            from sklearn.ensemble import RandomForestRegressor
            from sklearn.preprocessing import StandardScaler
            
            self.scaler = StandardScaler()
            X_scaled = self.scaler.fit_transform(X.reshape(X.shape[0], -1))
            
            self.model = RandomForestRegressor(n_estimators=100, random_state=42)
            self.model.fit(X_scaled, y)
            
            # Save model
            os.makedirs("models", exist_ok=True)
            joblib.dump(self.model, self.model_path)
            joblib.dump(self.scaler, self.scaler_path)
            
            logger.info("Model trained and saved successfully")
            
        except Exception as e:
            logger.error(f"Error training model: {e}")

    def _create_sequences(self, features: np.ndarray, targets: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
        """Create sequences for LSTM training"""
        X, y = [], []
        
        for i in range(self.sequence_length, len(features)):
            X.append(features[i-self.sequence_length:i])
            y.append(targets[i])
        
        return np.array(X), np.array(y)
