backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
psycopg[binary]==3.2.10
geoalchemy2==0.15.2
alembic==1.13.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.9.0
httpx==0.27.2
jinja2==3.1.4
xhtml2pdf==0.2.15
email-validator==2.2.0
openai==1.47.0
# Rain prediction dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2


